fixed tests

+++ compute/cloud-client/src/test/java/compute/Util.java
@@ -16,6 +16,7 @@
 
 package compute;
 
+import com.google.cloud.compute.v1.AggregatedListInstancesRequest;
 import com.google.cloud.compute.v1.Disk;
 import com.google.cloud.compute.v1.DisksClient;
 import com.google.cloud.compute.v1.Instance;
@@ -23,15 +24,11 @@ import com.google.cloud.compute.v1.InstanceTemplate;
 import com.google.cloud.compute.v1.InstanceTemplatesClient;
 import com.google.cloud.compute.v1.InstancesClient;
 import com.google.cloud.compute.v1.RegionDisksClient;
-import com.google.cloud.compute.v1.Reservation;
-import com.google.cloud.compute.v1.ReservationsClient;
-import com.google.cloud.compute.v1.Snapshot;
-import com.google.cloud.compute.v1.SnapshotsClient;
-import compute.deleteprotection.GetDeleteProtection;
-import compute.deleteprotection.SetDeleteProtection;
-import compute.disks.DeleteDisk;
-import compute.disks.DeleteSnapshot;
-import compute.reservation.DeleteReservation;
+import com.google.cloud.compute.v1.InstancesClient.AggregatedListPagedResponse;
+import com.google.cloud.compute.v1.InstancesScopedList;
+import com.google.cloud.compute.v1.ListInstanceTemplatesRequest;
+import com.google.cloud.compute.v1.InstanceTemplatesClient.ListPagedResponse;
+import com.google.cloud.compute.v1.Instance.Status;
 import java.io.IOException;
 import java.nio.charset.StandardCharsets;
 import java.security.SecureRandom;
@@ -39,6 +36,7 @@ import java.time.Instant;
 import java.time.OffsetDateTime;
 import java.time.temporal.ChronoUnit;
 import java.util.Base64;
+import java.util.Map;
 import java.util.Random;
 import java.util.concurrent.ExecutionException;
 import java.util.concurrent.TimeoutException;
@@ -59,41 +57,65 @@ public abstract class Util {
   // has creation timestamp >24 hours.
   public static void cleanUpExistingInstanceTemplates(String prefixToDelete, String projectId)
       throws IOException, ExecutionException, InterruptedException, TimeoutException {
-    try (InstanceTemplatesClient client = InstanceTemplatesClient.create()) {
-      for (InstanceTemplate template : client.list(projectId).iterateAll()) {
-        if (!template.hasCreationTimestamp()) {
-          continue;
-        }
-        if (template.getName().contains(prefixToDelete)
-            && isCreatedBeforeThresholdTime(template.getCreationTimestamp())
-            && template.isInitialized()) {
-          DeleteInstanceTemplate.deleteInstanceTemplate(projectId, template.getName());
+    for (InstanceTemplate template : listFilteredInstanceTemplates(projectId, prefixToDelete)
+        .iterateAll()) {
+      if (!template.hasCreationTimestamp()) {
+        continue;
+      }
+      if (template.getName().contains(prefixToDelete)
+          && isCreatedBeforeThresholdTime(template.getCreationTimestamp())
+          && template.isInitialized()) {
+        DeleteInstanceTemplate.deleteInstanceTemplate(projectId, template.getName());
         }
       }
     }
-  }
 
   // Delete instances which starts with the given prefixToDelete and
   // has creation timestamp >24 hours.
   public static void cleanUpExistingInstances(String prefixToDelete, String projectId,
       String instanceZone)
       throws IOException, ExecutionException, InterruptedException, TimeoutException {
-    try (InstancesClient instancesClient = InstancesClient.create()) {
-      for (Instance instance : instancesClient.list(projectId, instanceZone).iterateAll()) {
+    for (Map.Entry<String, InstancesScopedList> instanceGroup : listFilteredInstances(
+        projectId, prefixToDelete).iterateAll()) {
+      for (Instance instance : instanceGroup.getValue().getInstancesList()) {
         if (!instance.hasCreationTimestamp()) {
           continue;
         }
-        if (GetDeleteProtection.getDeleteProtection(projectId, instanceZone, instance.getName())) {
-          SetDeleteProtection.setDeleteProtection(
-              projectId, instanceZone, instance.getName(), false);
-        }
         if (instance.getName().contains(prefixToDelete)
-            && isCreatedBeforeThresholdTime(instance.getCreationTimestamp())) {
+            && isCreatedBeforeThresholdTime(instance.getCreationTimestamp())
+            && instance.getStatus().equalsIgnoreCase(Status.RUNNING.toString())) {
           DeleteInstance.deleteInstance(projectId, instanceZone, instance.getName());
         }
       }
     }
   }
+public static AggregatedListPagedResponse listFilteredInstances(String project,
+                                                                String instanceNamePrefix) throws IOException {
+  try (InstancesClient instancesClient = InstancesClient.create()) {
+
+    AggregatedListInstancesRequest aggregatedListInstancesRequest = AggregatedListInstancesRequest
+        .newBuilder()
+        .setProject(project)
+        .setFilter(String.format("name:%s", instanceNamePrefix))
+        .build();
+
+    return instancesClient
+        .aggregatedList(aggregatedListInstancesRequest);
+  }
+}
+
+public static ListPagedResponse listFilteredInstanceTemplates(String projectId,
+                                                              String instanceTemplatePrefix) throws IOException {
+  try (InstanceTemplatesClient instanceTemplatesClient = InstanceTemplatesClient.create()) {
+    ListInstanceTemplatesRequest listInstanceTemplatesRequest =
+        ListInstanceTemplatesRequest.newBuilder()
+            .setProject(projectId)
+            .setFilter(String.format("name:%s", instanceTemplatePrefix))
+            .build();
+
+    return instanceTemplatesClient.list(listInstanceTemplatesRequest);
+  }
+}
 
   public static boolean isCreatedBeforeThresholdTime(String timestamp) {
     return OffsetDateTime.parse(timestamp).toInstant()
@@ -156,54 +178,5 @@ public abstract class Util {
       return defaultValue;
     }
     return val;
-  }
-
-  // Delete reservations which starts with the given prefixToDelete and
-  // has creation timestamp >24 hours.
-  public static void cleanUpExistingReservations(String prefixToDelete, String projectId,
-                                                 String zone)
-      throws IOException, ExecutionException, InterruptedException, TimeoutException {
-    try (ReservationsClient reservationsClient = ReservationsClient.create()) {
-      for (Reservation reservation : reservationsClient.list(projectId, zone).iterateAll()) {
-        if (!reservationsClient.list(projectId, zone).iterateAll().iterator().hasNext()) {
-          break;
-        }
-        if (reservation.getName().contains(prefixToDelete)
-            && isCreatedBeforeThresholdTime(reservation.getCreationTimestamp())) {
-          DeleteReservation.deleteReservation(projectId, zone, reservation.getName());
-        }
-      }
-    }
-  }
-
-  public static void cleanUpExistingDisks(String prefixToDelete, String projectId,
-                                                 String zone)
-      throws IOException, ExecutionException, InterruptedException, TimeoutException {
-    try (DisksClient disksClient = DisksClient.create()) {
-      for (Disk disk : disksClient.list(projectId, zone).iterateAll()) {
-        if (!disksClient.list(projectId, zone).iterateAll().iterator().hasNext()) {
-          break;
-        }
-        if (disk.getName().contains(prefixToDelete)
-            && isCreatedBeforeThresholdTime(disk.getCreationTimestamp())) {
-          DeleteDisk.deleteDisk(projectId, zone, disk.getName());
-        }
-      }
-    }
-  }
-
-  public static void cleanUpExistingSnapshots(String prefixToDelete, String projectId)
-      throws IOException, ExecutionException, InterruptedException, TimeoutException {
-    try (SnapshotsClient snapshotsClient = SnapshotsClient.create()) {
-      for (Snapshot snapshot : snapshotsClient.list(projectId).iterateAll()) {
-        if (!snapshotsClient.list(projectId).iterateAll().iterator().hasNext()) {
-          break;
-        }
-        if (snapshot.getName().contains(prefixToDelete)
-            && isCreatedBeforeThresholdTime(snapshot.getCreationTimestamp())) {
-          DeleteSnapshot.deleteSnapshot(projectId, snapshot.getName());
-        }
-      }
-    }
   }
 }
